refactor(mfa): share the device eligibility check between the actor and the 3DS hook
The 3DS navigation hook re-implemented the validateDevice gates inline, and in the
worse order: it queried the platform capability before the cheap allowed-methods
check. The actor body now lives in checkDeviceEligibility next to the operations
module, the actor only wraps it in fromPromise, and the hook reads the result's
success flag and logs the granular refusal reason.
